Create read::any module
